requirement.txt
# Core dependencies
openai>=1.12.0              # 或者使用其他LLM API
anthropic>=0.18.0           # 如果使用Claude
langchain>=0.1.0
langchain-community>=0.0.20

# Database connectors
sqlalchemy>=2.0.0
pymysql>=1.1.0              # MySQL
psycopg2-binary>=2.9.0      # PostgreSQL
pyodbc>=5.0.0               # SQL Server

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web interface
streamlit>=1.31.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0